*: Fix clippy warnings (#2139)
* Fix needless question mark operator

* Don't convert from u64 to u64

LocalStreamId is already a u64, no need to convert.

* Don't use `.into()` to convert to the same type

* Don't specify lifetime if it can be inferred

* Use `vec!` macro if we immediately push to it

This creates the vector with the appropriate capacity.

* Don't index array when taking a reference is enough

+++ muxers/mplex/src/codec.rs
@@ -244,25 +244,25 @@ impl Encoder for Codec {
     fn encode(&mut self, item: Self::Item, dst: &mut BytesMut) -> Result<(), Self::Error> {
         let (header, data) = match item {
             Frame::Open { stream_id } => {
-                (u64::from(stream_id.num) << 3, Bytes::new())
+                (stream_id.num << 3, Bytes::new())
             },
             Frame::Data { stream_id: LocalStreamId { num, role: Endpoint::Listener }, data } => {
-                (u64::from(num) << 3 | 1, data)
+                (num << 3 | 1, data)
             },
             Frame::Data { stream_id: LocalStreamId { num, role: Endpoint::Dialer }, data } => {
-                (u64::from(num) << 3 | 2, data)
+                (num << 3 | 2, data)
             },
             Frame::Close { stream_id: LocalStreamId { num, role: Endpoint::Listener } } => {
-                (u64::from(num) << 3 | 3, Bytes::new())
+                (num << 3 | 3, Bytes::new())
             },
             Frame::Close { stream_id: LocalStreamId { num, role: Endpoint::Dialer } } => {
-                (u64::from(num) << 3 | 4, Bytes::new())
+                (num << 3 | 4, Bytes::new())
             },
             Frame::Reset { stream_id: LocalStreamId { num, role: Endpoint::Listener } } => {
-                (u64::from(num) << 3 | 5, Bytes::new())
+                (num << 3 | 5, Bytes::new())
             },
             Frame::Reset { stream_id: LocalStreamId { num, role: Endpoint::Dialer } } => {
-                (u64::from(num) << 3 | 6, Bytes::new())
+                (num << 3 | 6, Bytes::new())
             },
         };
